+++ backend/src/main/java/com/ll/cafeservice/global/exception/GlobalExceptionHandler.java
@@ -2,6 +2,8 @@ package com.ll.cafeservice.global.exception;
 
 import com.ll.cafeservice.api.Empty;
 import com.ll.cafeservice.api.Result;
+import com.ll.cafeservice.security.exception.CustomJwtException;
+import com.ll.cafeservice.security.exception.InvalidCredentialsException;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
