+++ backend/src/main/java/com/ll/cafeservice/security/handler/CustomAuthenticationEntryPoint.java
@@ -1,4 +1,4 @@
-package com.ll.cafeservice.global.exception;
+package com.ll.cafeservice.security.handler;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
