+++ backend/src/main/java/com/ll/cafeservice/security/jwt/JwtAuthenticationFilter.java
@@ -1,6 +1,6 @@
-package com.ll.cafeservice.global.jwt;
+package com.ll.cafeservice.security.jwt;
 
-import com.ll.cafeservice.global.exception.CustomJwtException;
+import com.ll.cafeservice.security.exception.CustomJwtException;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
